# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/comments": {
            "post": {
                "description": "Создает новый комментарий к фото",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Создать комментарий",
                "parameters": [
                    {
                        "description": "Данные комментария",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: comment created successfully, id: 1",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный ввод",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/comments/{id}/delete": {
            "delete": {
                "description": "Удаляет комментарий по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Удалить комментарий",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID комментария",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: comment deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/comments/{id}/edit": {
            "put": {
                "description": "Обновляет текст комментария",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Обновить комментарий",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID комментария",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные комментария",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: comment updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/comments/{photoID}": {
            "get": {
                "description": "Возвращает список комментариев по photo_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Получить комментарии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фото",
                        "name": "photoID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный photo_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Проверяет учетные данные пользователя и выдает JWT-токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные пользователя",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token: JWT токен, username: Имя пользователя",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ввод",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка генерации токена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создает нового пользователя и отправляет подтверждение по email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: Успешная регистрация. Пожалуйста подтвердите email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ввод",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Текст комментария",
                    "type": "string",
                    "example": "Отличное фото!"
                },
                "created_at": {
                    "description": "Дата создания комментария",
                    "type": "string",
                    "example": "2024-01-31T12:45:00Z"
                },
                "id": {
                    "description": "ID комментария",
                    "type": "integer",
                    "example": 1
                },
                "photo_id": {
                    "description": "ID фото, к которому относится комментарий",
                    "type": "integer",
                    "example": 101
                },
                "updated_at": {
                    "description": "Дата последнего обновления комментария",
                    "type": "string",
                    "example": "2024-01-31T12:50:00Z"
                },
                "user_id": {
                    "description": "ID пользователя, оставившего комментарий",
                    "type": "integer",
                    "example": 42
                },
                "username": {
                    "description": "Имя пользователя, оставившего комментарий",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email пользователя",
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "description": "ID пользователя",
                    "type": "integer",
                    "example": 1
                },
                "password": {
                    "description": "Пароль пользователя (не возвращается в ответах)",
                    "type": "string",
                    "example": "securepassword"
                },
                "username": {
                    "description": "Имя пользователя",
                    "type": "string",
                    "example": "johndoe"
                },
                "verified": {
                    "description": "Флаг подтверждения email",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}